requirements.txt
requests
pandas
matplotlib
supabase
python-dotenv
runpod